chore: move authErrorCategory functions from utils to router/transformer
